Refatoração
JSONHandler agora possui método para converter JSONObjects para a classe
Bill.

Adicionado teste da classe JSONHandler

+++ app/src/main/java/com/nubank/allan/billscreen/model/LineItem.java
@@ -11,12 +11,12 @@ public class LineItem {
     private String href;
     private int index;
     private int charges;
-    private double amount;
+    private int amount;
 
     public LineItem() {
     }
 
-    public LineItem(Date postDate, String title, String href, int index, int charges, double amount) {
+    public LineItem(Date postDate, String title, String href, int index, int charges, int amount) {
         this.postDate = postDate;
         this.title = title;
         this.href = href;
@@ -65,11 +65,11 @@ public class LineItem {
         this.charges = charges;
     }
 
-    public double getAmount() {
+    public int getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(int amount) {
         this.amount = amount;
     }
 }
