+++ app/src/main/java/com/nubank/allan/billscreen/model/Summary.java
@@ -9,18 +9,18 @@ public class Summary {
 
     private Date dueDate;
     private Date closeDate;
-    private double pastBalance;
-    private double totalBalance;
-    private double interest;
-    private double totalCumulative;
-    private double paid;
-    private double minPayment;
+    private int pastBalance;
+    private int totalBalance;
+    private int interest;
+    private int totalCumulative;
+    private int paid;
+    private int minPayment;
     private Date openDate;
 
     public Summary() {
     }
 
-    public Summary(Date dueDate, Date closeDate, double pastBalance, double totalBalance, double interest, double totalCumulative, double paid, double minPayment, Date openDate) {
+    public Summary(Date dueDate, Date closeDate, int pastBalance, int totalBalance, int interest, int totalCumulative, int paid, int minPayment, Date openDate) {
         this.dueDate = dueDate;
         this.closeDate = closeDate;
         this.pastBalance = pastBalance;
@@ -48,51 +48,47 @@ public class Summary {
         this.closeDate = closeDate;
     }
 
-    public double getPastBalance() {
+    public int getPastBalance() {
         return pastBalance;
     }
 
-    public void setPastBalance(double pastBalance) {
+    public void setPastBalance(int pastBalance) {
         this.pastBalance = pastBalance;
     }
 
-    public double getTotalBalance() {
+    public int getTotalBalance() {
         return totalBalance;
     }
 
-    public void setTotalBalance(double totalBalance) {
+    public void setTotalBalance(int totalBalance) {
         this.totalBalance = totalBalance;
     }
 
-    public double getInterest() {
+    public int getInterest() {
         return interest;
     }
 
-    public void setInterest(double interest) {
+    public void setInterest(int interest) {
         this.interest = interest;
     }
 
-    public double getTotalCumulative() {
-        return totalCumulative;
-    }
+    public int getTotalCumulative() { return totalCumulative; }
 
-    public void setTotalCumulative(double totalCumulative) {
-        this.totalCumulative = totalCumulative;
-    }
+    public void setTotalCumulative(int totalCumulative) { this.totalCumulative = totalCumulative; }
 
-    public double getPaid() {
+    public int getPaid() {
         return paid;
     }
 
-    public void setPaid(double paid) {
+    public void setPaid(int paid) {
         this.paid = paid;
     }
 
-    public double getMinPayment() {
+    public int getMinPayment() {
         return minPayment;
     }
 
-    public void setMinPayment(double minPayment) {
+    public void setMinPayment(int minPayment) {
         this.minPayment = minPayment;
     }
